+++ src/components/Login/Login.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { auth } from '../../firebase';
+import { auth } from '../../firebase/firebase';
 import './Login.scss';
 
 export default function Login() {
